Conflicts:
	src/server/game/Accounts/RBAC.h
	src/server/game/Battlefield/Battlefield.cpp
	src/server/game/Battlegrounds/Zones/BattlegroundIC.h
	src/server/game/Battlegrounds/Zones/BattlegroundSA.cpp
	src/server/game/Battlegrounds/Zones/BattlegroundSA.h
	src/server/game/CMakeLists.txt
	src/server/game/Entities/Unit/Unit.cpp
	src/server/game/Quests/QuestDef.cpp
	src/server/game/Server/WorldSocket.cpp
	src/server/scripts/CMakeLists.txt
	src/server/scripts/EasternKingdoms/ScarletEnclave/chapter1.cpp
	src/server/scripts/EasternKingdoms/ScarletEnclave/chapter5.cpp
	src/server/scripts/EasternKingdoms/ScarletMonastery/boss_headless_horseman.cpp
	src/server/scripts/Kalimdor/zone_durotar.cpp
	src/server/scripts/Northrend/AzjolNerub/Ahnkahet/boss_jedoga_shadowseeker.cpp
	src/server/scripts/Northrend/AzjolNerub/Ahnkahet/boss_prince_taldaram.cpp
	src/server/scripts/Northrend/ChamberOfAspects/ObsidianSanctum/boss_sartharion.cpp
	src/server/scripts/Northrend/ChamberOfAspects/ObsidianSanctum/obsidian_sanctum.cpp
	src/server/scripts/Northrend/ChamberOfAspects/RubySanctum/boss_halion.cpp
	src/server/scripts/Northrend/CrusadersColiseum/TrialOfTheChampion/instance_trial_of_the_champion.cpp
	src/server/scripts/Northrend/CrusadersColiseum/TrialOfTheChampion/trial_of_the_champion.cpp
	src/server/scripts/Northrend/CrusadersColiseum/TrialOfTheCrusader/boss_anubarak_trial.cpp
	src/server/scripts/Northrend/CrusadersColiseum/TrialOfTheCrusader/boss_twin_valkyr.cpp
	src/server/scripts/Northrend/CrusadersColiseum/TrialOfTheCrusader/trial_of_the_crusader.h
	src/server/scripts/Northrend/FrozenHalls/ForgeOfSouls/boss_devourer_of_souls.cpp
	src/server/scripts/Northrend/IcecrownCitadel/boss_icecrown_gunship_battle.cpp
	src/server/scripts/Northrend/IcecrownCitadel/boss_lord_marrowgar.cpp
	src/server/scripts/Northrend/Nexus/EyeOfEternity/boss_malygos.cpp
	src/server/scripts/Northrend/Nexus/Oculus/boss_varos.cpp
	src/server/scripts/Northrend/Ulduar/HallsOfLightning/boss_ionar.cpp
	src/server/scripts/Northrend/Ulduar/HallsOfLightning/boss_loken.cpp
	src/server/scripts/Northrend/Ulduar/HallsOfLightning/boss_volkhan.cpp
	src/server/scripts/Northrend/Ulduar/HallsOfStone/halls_of_stone.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_assembly_of_iron.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_general_vezax.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_hodir.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_kologarn.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_mimiron.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_razorscale.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_xt002.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_yogg_saron.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/instance_ulduar.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/ulduar.h
	src/server/scripts/Northrend/UtgardeKeep/UtgardePinnacle/boss_ymiron.cpp
	src/server/scripts/Northrend/VioletHold/violet_hold.cpp
	src/server/scripts/Northrend/zone_grizzly_hills.cpp
	src/server/scripts/Northrend/zone_sholazar_basin.cpp
	src/server/scripts/Northrend/zone_zuldrak.cpp
	src/server/scripts/Spells/spell_shaman.cpp

+++ src/server/shared/Utilities/Util.h
@@ -616,7 +616,7 @@ class EventMap
             if (!phase)
                 _phase = 0;
             else if (phase <= 8)
-                _phase = (1 << (phase - 1));
+                _phase = uint8(1 << (phase - 1));
         }
 
         /**
@@ -627,7 +627,7 @@ class EventMap
         void AddPhase(uint8 phase)
         {
             if (phase && phase <= 8)
-                _phase |= (1 << (phase - 1));
+                _phase |= uint8(1 << (phase - 1));
         }
 
         /**
@@ -638,7 +638,7 @@ class EventMap
         void RemovePhase(uint8 phase)
         {
             if (phase && phase <= 8)
-                _phase &= ~(1 << (phase - 1));
+                _phase &= uint8(~(1 << (phase - 1)));
         }
 
         /**
